refactor : APIType file 분리 (#336)
