+++ KCS/KCS/Data/Network/APIType.swift
@@ -1,0 +1,19 @@
+//
+//  APIType.swift
+//  KCS
+//
+//  Created by 김영현 on 2/25/24.
+//
+
+import Foundation
+
+enum APIType {
+    
+    case getStores
+    case getImage
+    case getSearchStores
+    case getAutoCompletion
+    case postNewStoreRequest
+    case storeUpdateRequest
+    
+}
